python/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.6
requests==2.24.0
rope==0.18.0
urllib3==1.25.11
